Refactors FulltextSortType into the ongdb-kernel module and 'org.neo4j.kernel.impl.util' package.
