+++ community/fulltext-index/src/main/java/org/neo4j/kernel/api/impl/fulltext/LuceneFulltextDocumentStructure.java
@@ -44,6 +44,7 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.Map;
 
+import org.neo4j.kernel.impl.util.FulltextSortType;
 import org.neo4j.values.storable.TextValue;
 import org.neo4j.values.storable.Value;
 import org.neo4j.values.storable.ValueGroup;
