+++ community/kernel/src/main/java/org/neo4j/kernel/impl/util/FulltextSortType.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.api.impl.fulltext;
+package org.neo4j.kernel.impl.util;
 
 public enum FulltextSortType
 {
@@ -56,5 +56,10 @@ public enum FulltextSortType
         default:
             return null;
         }
+    }
+
+    public int getNeoStoreByte()
+    {
+        return neoStoreByte;
     }
 }
